# lmbench.py: Fixing Compilation errors