1.Added libtirpc-devel package for rpc related compilation error
2.Added extra args to make command for pmap related compilation
errors
3.Added patch file in order to get rid of lsleek related
compilation error

Signed-off-by: Shirisha Ganta <SHIRISHA.Ganta1@ibm.com>