@shirishaganta1 2021-04-19
Before patch:
# avocado run lmbench.py
Fetching asset from lmbench.py:Lmbench.test
JOB ID     : 28196bbb36fced28adb1d512a0a808f7d8d64476
JOB LOG    : /home/siri/avocado-fvt-wrapper/results/job-2021-04-18T23.42-28196bb/job.log
 (1/1) lmbench.py:Lmbench.test: ERROR: Command 'make' failed.\nstdout: b'cd src && make\nmake[1]: Entering directory \'/var/tmp/avocado_rvdhkh37/avocado_job_uez_keti/1-lmbench.py_Lmbench.test/lmbench3/src\'\ngmake[2]: Entering directory \'/var/tmp/avocado_rvdhkh37/avocado_job_uez_keti/1-lmbench... (3.19 s)
RESULTS    : PASS 0 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/siri/avocado-fvt-wrapper/results/job-2021-04-18T23.42-28196bb/results.html
JOB TIME   : 4.39 s
After Patch:
# avocado run lmbench.py
Fetching asset from lmbench.py:Lmbench.test
JOB ID     : 7545ace10a7a1452ae1d1c9cb34b4fd99a910649
JOB LOG    : /home/siri/avocado-fvt-wrapper/results/job-2021-04-16T18.47-7545ace/job.log
 (1/1) lmbench.py:Lmbench.test: /Using config in CONFIG.localhost
Friday 16 April 2021 06:47:49 PM IST
Latency measurements
\Friday 16 April 2021 06:48:23 PM IST
Calculating file system latency
|Friday 16 April 2021 06:48:24 PM IST
Local networking
|Friday 16 April 2021 06:52:11 PM IST
Bandwidth measurements
/Friday 16 April 2021 06:53:24 PM IST
Calculating context switch overhead
/Friday 16 April 2021 06:53:28 PM IST
Calculating effective TLB size
-Friday 16 April 2021 06:53:29 PM IST
Calculating memory load parallelism
\Friday 16 April 2021 06:54:02 PM IST
McCalpin's STREAM benchmark
/Friday 16 April 2021 06:54:04 PM IST
Calculating memory load latency
/
\Friday 16 April 2021 07:13:44 PM IST
/Using config in CONFIG.localhost
Friday 16 April 2021 07:13:45 PM IST
Latency measurements
/Friday 16 April 2021 07:14:18 PM IST
Calculating file system latency
-Friday 16 April 2021 07:14:19 PM IST
Local networking
|Friday 16 April 2021 07:18:07 PM IST
Bandwidth measurements
-
\Friday 16 April 2021 07:19:18 PM IST
Calculating context switch overhead
\Friday 16 April 2021 07:19:22 PM IST
Calculating effective TLB size
|Friday 16 April 2021 07:19:23 PM IST
Calculating memory load parallelism
-Friday 16 April 2021 07:19:57 PM IST
McCalpin's STREAM benchmark
\Friday 16 April 2021 07:19:58 PM IST
Calculating memory load latency
\
/Friday 16 April 2021 07:39:37 PM IST
PASS (3125.46 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/siri/avocado-fvt-wrapper/results/job-2021-04-16T18.47-7545ace/results.html
JOB TIME   : 3126.04 s